+++ src/libs/antares/study/CMakeLists.txt
@@ -102,9 +102,9 @@ set(SRC_STUDY_PART_SHORT_TERM_STORAGE
         parts/short-term-storage/series.cpp
         include/antares/study/parts/short-term-storage/series.h
         include/antares/study/parts/short-term-storage/cluster.h
-        include/antares/study/parts/short-term-storage/AdditionalConstraints.h
+        include/antares/study/parts/short-term-storage/additionalConstraints.h
         parts/short-term-storage/cluster.cpp
-        parts/short-term-storage/AdditionalConstraints.cpp
+        parts/short-term-storage/additionalConstraints.cpp
 )
 source_group("study\\part\\short-term-storage" FILES ${SRC_STUDY_PART_SHORT_TERM_SOTRAGE})
