+++ src/libs/antares/study/include/antares/study/parts/short-term-storage/additionalConstraints.h
