+++ src/libs/antares/study/include/antares/study/parts/short-term-storage/cluster.h
@@ -26,7 +26,7 @@
 
 #include <antares/inifile/inifile.h>
 
-#include "AdditionalConstraints.h"
+#include "additionalConstraints.h"
 #include "properties.h"
 #include "series.h"
