+++ src/libs/antares/study/parts/short-term-storage/additionalConstraints.cpp
@@ -19,7 +19,7 @@
 ** along with Antares_Simulator. If not, see <https://opensource.org/license/mpl-2-0/>.
 */
 
-#include "antares/study/parts/short-term-storage/AdditionalConstraints.h"
+#include "antares/study/parts/short-term-storage/additionalConstraints.h"
 
 #include <algorithm>
